requirements.txt
comtypes
litellm
pycaw
python-dotenv
python-docx
python-pptx
customtkinter
requests
pywin32
pillow
google-genai
pyautogui
replicate